# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Habitat Commissioner",
    "description": "Express REST API for moving data between orgs and environments in apigee",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/kvms/apigee/list": {
      "get": {
        "description": "List all KVMs in the specified environment in Apigee",
        "summary": "List KVMs",
        "tags": [
          "kvms"
        ],
        "operationId": "listKvms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/apigee/list/{kvm_name}": {
      "get": {
        "description": "Get a list of all the entries for a specified KVM in Apigee",
        "summary": "List KVM entries",
        "tags": [
          "kvms"
        ],
        "operationId": "listKvmEntries",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/apigee/details/{kvm_name}": {
      "get": {
        "description": "Details of a single KVM in Apigee",
        "summary": "Get KVM",
        "tags": [
          "kvms"
        ],
        "operationId": "getKVM",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/apigee/details/{kvm_name}/{entry_name}": {
      "get": {
        "description": "Get the details of a single KVM entry in Apigee",
        "summary": "Get KVM entry",
        "tags": [
          "kvms"
        ],
        "operationId": "getKvmEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          },
          {
            "name": "entry_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "entry name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/repo/list": {
      "get": {
        "description": "List of KVMs in the specified environment from the specified repo",
        "summary": "List KVMs in the repo",
        "tags": [
          "kvms"
        ],
        "operationId": "listRepoKvms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/repo/list/{kvm_name}": {
      "get": {
        "description": "List of all entries for a single KVM in the repo",
        "summary": "List KVM entries for repo KVM",
        "tags": [
          "kvms"
        ],
        "operationId": "listRepoKvmEntries",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/repo/details/{kvm_name}": {
      "get": {
        "description": "Details of KVM in the repo",
        "summary": "Get repo KVM",
        "tags": [
          "kvms"
        ],
        "operationId": "getRepoKvm",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/repo/details/{kvm_name}/{entry_name}": {
      "get": {
        "description": "Get the details for a single KVM entry in the repo",
        "summary": "Get KVM entry for a KVM from repo",
        "tags": [
          "kvms"
        ],
        "operationId": "getRepoKvmEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          },
          {
            "name": "entry_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "entry name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/apigee": {
      "post": {
        "description": "Export KVMs from the repo into Apigee",
        "summary": "Import KVMs",
        "tags": [
          "kvms"
        ],
        "operationId": "importKvms",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/apigee/{kvm_name}": {
      "put": {
        "description": "Update an existing KVM entry using data from the repo",
        "summary": "Update a KVM entry",
        "tags": [
          "kvms"
        ],
        "operationId": "updateKvmEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      },
      "post": {
        "description": "Create new entries in an existing KVM in Apigee using data from the repo",
        "summary": "Create a new KVM entry",
        "tags": [
          "kvms"
        ],
        "operationId": "createKvmEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/{kvm_name}": {
      "delete": {
        "description": "Delete a KVM in Apigee",
        "summary": "Delete KVM",
        "tags": [
          "kvms"
        ],
        "operationId": "deleteKvm",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/kvms/{kvm_name}/{entry_name}": {
      "delete": {
        "description": "Delete a KVM entry from an existing KVM in Apigee",
        "summary": "Delete KVM entry",
        "tags": [
          "kvms"
        ],
        "operationId": "deleteKvmEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "kvm_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "KVM name"
          },
          {
            "name": "entry_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Entry name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApigeeResponse": {
      "title": "apigeeResponse",
      "type": "object",
      "properties": {
        "code": {
          "description": "",
          "example": 200,
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "description": "",
          "example": "Context sensitive message",
          "type": "string"
        },
        "data": {
          "description": "Generally reflects the response directly from the Apigee management API. May be an object or an array",
          "type": "object"
        }
      }
    },
    "ApigeeRequest": {
      "type": "array",
      "items": {
        "type": "string",
        "description": "item name or ID",
        "example": "item-name-or-id"
      }
    },
...
  }
}